+++ app/src/main/java/org/android/go/sopt/playlist/PlaylistTitleAdapter.kt
@@ -2,7 +2,6 @@ package org.android.go.sopt.playlist
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import org.android.go.sopt.databinding.ItemPlaylistTitleBinding
 
@@ -19,15 +18,5 @@ class PlaylistTitleAdapter :
 
     override fun onBindViewHolder(holder: PlaylistTitleViewHolder, position: Int) {
         holder.onBind(getItem(position))
-    }
-}
-
-object PlaylistTitleDiffCallback : DiffUtil.ItemCallback<PlaylistTitle>() {
-    override fun areItemsTheSame(oldItem: PlaylistTitle, newItem: PlaylistTitle): Boolean {
-        return oldItem.id == newItem.id
-    }
-
-    override fun areContentsTheSame(oldItem: PlaylistTitle, newItem: PlaylistTitle): Boolean {
-        return oldItem == newItem
     }
 }
