+++ app/src/main/java/org/android/go/sopt/playlist/PlaylistTitleViewHolder.kt
@@ -3,7 +3,7 @@ package org.android.go.sopt.playlist
 import androidx.recyclerview.widget.RecyclerView
 import org.android.go.sopt.databinding.ItemPlaylistTitleBinding
 
-class PlaylistTitleViewHolder (private val binding: ItemPlaylistTitleBinding) :
+class PlaylistTitleViewHolder(private val binding: ItemPlaylistTitleBinding) :
     RecyclerView.ViewHolder(binding.root) {
     fun onBind(item: PlaylistTitle) {
         binding.tvPlaylistTitle.text = item.title
